Restructured to break out recurssion into its own file.
